+++ tests/core/c/src/x_ram_checks.c
@@ -495,7 +495,7 @@ uint32_t X_RAM_CHECKS_checkAddress16(void)
 	
 	for (z = 0; z < X_RAM_CHECKS_get16bitZoneNumber(); z++)
 	{
-    while (p < (((zones[z].end_address - zones[z].start_address) / (sizeof(uint16_t))) * (UINT16_MAX + 1)))
+    while (p < ((zones[z].end_address - zones[z].start_address) / (sizeof(uint16_t) * (UINT16_MAX + 1))))
     {
       pointer = (uint16_t *) zones[z].start_address;  
 			for (i = 0; i < UINT16_MAX; i++)